Add jitter to Force layout
ForceLayout.java: Add some jitter, to prevent things like 2 nodes that
really should be pulled past each other getting stuck cause they're /just/
so in a position that the forces are perfectly balanced.  Jiggle the
attractive force on lines a little bit, and add some slight entropy to
direction of resultant velocity to allow such situations to become unstuck
and to generally add some entropy to the algorithm to allow it to escape
from narrow local minima.

Make the minkve and jitter parameters setable.

Note that this means the the algorithm is unlikely to go cold and stop by
itself. Users will have to tweak jitter and minkve.

Warning: The minkve is against the sum of all kves at present. This is
probably not useful, as it means a large graph with nodes almost rest could
have a larger sum-kve than a small graph with higher energy nodes.  Will
probably change this in future to test against the maximum kve of all nodes
rather than the sum.